+++ core/defense/md_dnn.py
@@ -109,7 +109,7 @@ class DNNMalwareDetector(nn.Module):
         self.eval()
         with torch.no_grad():
             for x, y in test_data_producer:
-                x, y = utils.to_device(x.float(), y.long(), self.device)
+                x, y = utils.to_device(x.double(), y.long(), self.device)
                 logits = self.forward(x)
                 confidences.append(F.softmax(logits, dim=-1))
                 gt_labels.append(y)
@@ -130,9 +130,9 @@ class DNNMalwareDetector(nn.Module):
         ig = IntegratedGradients(_ig_wrapper)
 
         for i, (x, y) in enumerate(test_data_producer):
-            x, y = utils.to_device(x.float(), y.long(), self.device)
+            x, y = utils.to_device(x.double(), y.long(), self.device)
             x.requires_grad = True
-            baseline = torch.zeros_like(x, dtype=torch.float32, device=self.device)
+            baseline = torch.zeros_like(x, dtype=torch.double, device=self.device)
             attribution_bs = ig.attribute(x,
                                           baselines=baseline,
                                           target=target_label)
@@ -208,7 +208,7 @@ class DNNMalwareDetector(nn.Module):
             self.train()
             losses, accuracies = [], []
             for idx_batch, (x_train, y_train) in enumerate(train_data_producer):
-                x_train, y_train = utils.to_device(x_train.float(), y_train.long(), self.device)
+                x_train, y_train = utils.to_device(x_train.double(), y_train.long(), self.device)
                 start_time = time.time()
                 optimizer.zero_grad()
                 logits = self.forward(x_train)
@@ -231,7 +231,7 @@ class DNNMalwareDetector(nn.Module):
             avg_acc_val = []
             with torch.no_grad():
                 for x_val, y_val in validation_data_producer:
-                    x_val, y_val = utils.to_device(x_val.float(), y_val.long(), self.device)
+                    x_val, y_val = utils.to_device(x_val.double(), y_val.long(), self.device)
                     logits = self.forward(x_val)
                     acc_val = (logits.argmax(1) == y_val).sum().item()
                     acc_val /= x_val.size()[0]
